requirements-wheels.txt
pydantic == 1.10.*
PyYAML == 6.0
types-PyYAML == 6.0.*
requests == 2.28.*
types-requests == 2.28.*
setproctitle == 1.3.*
pynetstring == 0.5
